add pyvips
pyvips is a binding for the libvips image processing library. It's
fast and only needs a little memory. For example, on this benchmark:

https://github.com/jcupitt/libvips/wiki/Speed-and-memory-use

It's 3x faster than ImageMagick and needs 5x less memory.

pyvips works on all python versions on all platforms, is LGPL, can be
simply installed with pip, has complete documentation, has a large test
suite and has no memory leaks.